Handled data input to get predictions
